requirements.txt
# Core dependencies
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0

# LLM and NLP
llama-cpp-python>=0.2.20
openai>=1.0.0
tiktoken>=0.5.0
langchain>=0.0.350
langchain-community>=0.0.10

# Web interface and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0

# Data processing and utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
jsonlines>=4.0.0
pathlib>=1.0.1



# Development tools (optional)
# pytest>=7.4.0
# black>=23.10.0
# flake8>=6.1.0
# mypy>=1.7.0

# System utilities
psutil>=5.9.0
GPUtil>=1.4.0